Optimized list handling with `FlatList`
Now, what ScrollView does is, it always renders all the items that are inside of it whenever the overall UI is rendered. So even if an item is not currently on the screen like this last list item here, this item is still rendered behind the scenes even though it's not visible yet.
And therefore I will actually name this itemData, because it's actually an object that does not just contain our values, but also some metadata.
Now, with that, we can get rid of this key prop because we're no longer mapping this list manually and hence keys are no longer managed like this.
Instead when using FlatList, there are two main ways of adding keys to these list items. 1. The first approach is to turn your data values from primitive values like strings as we have it here, into objects that have a `key` property. 2. You also got an alternative to setting up such a key property in the input data. The `keyExtractor` prop wants a function as a value, for example, an inline arrow function.