Building custom buttons (by `Pressable`) & Deleting items
Bind is a standard JavaScript function which basically allows you pre-configure a function for future execution.
Now, the first value you pass to bind sets the this keyword in the to be executed function and we don't care about that here, so I'll just set it to this, but it doesn't matter.
The second value, the second argument you pass to bind will be the first parameter received by the to be called function.