Final touches & finishing the first app
But if we had multiple screens in this app, so not just the model, but also other screens, which we will have later once we add navigation, then manually setting the background color like this all the time, can become annoying.

Therefore what you can do when using `expo`, is you can go to the `app.json` file. And in that file, you can actually also add a special `backgroundColor` configuration item and set this to a hex code. And this will then automatically be applied as a background color to all screens by expo.

So now with the styling changed, there's one thing that might become obvious. The status bar is really difficult to see on both these devices. The reason for that is that it's still black, that it doesn't know that we have a dark background.

Now, thankfully, expo gives us a solution for that. We can import a special component from expo. And that component is the `StatusBar`. We can import it from `expo-status-bar`, which is a standalone package which is already part of this project. This is a component which we can use to fine tune the look of the status bar.